refactor class name and move to logs package
