+++ app/src/main/java/com/parishod/watomatic/model/logs/AutoReplyLogsDB.java
@@ -1,4 +1,4 @@
-package com.parishod.watomatic.model.logs.whatsapp;
+package com.parishod.watomatic.model.logs;
 
 import android.content.Context;
 
@@ -6,16 +6,18 @@ import androidx.room.Database;
 import androidx.room.Room;
 import androidx.room.RoomDatabase;
 
+import com.parishod.watomatic.model.logs.whatsapp.WhatsappAutoReplyLogs;
+import com.parishod.watomatic.model.logs.whatsapp.WhatsappAutoReplyLogsDao;
 import com.parishod.watomatic.model.utils.Constants;
 
 @Database(entities = {WhatsappAutoReplyLogs.class}, version = 1)
-public abstract class WhatsappAutoReplyLogsDB extends RoomDatabase {
+public abstract class AutoReplyLogsDB extends RoomDatabase {
     private static final String DB_NAME = Constants.WHATSAPP_LOGS_DB_NAME;
-    private static WhatsappAutoReplyLogsDB _instance;
+    private static AutoReplyLogsDB _instance;
 
-    public static synchronized WhatsappAutoReplyLogsDB getInstance(Context context){
+    public static synchronized AutoReplyLogsDB getInstance(Context context){
         if(_instance == null){
-            _instance = Room.databaseBuilder(context.getApplicationContext(), WhatsappAutoReplyLogsDB.class, DB_NAME)
+            _instance = Room.databaseBuilder(context.getApplicationContext(), AutoReplyLogsDB.class, DB_NAME)
             .fallbackToDestructiveMigration()
             .allowMainThreadQueries()
             .build();
